\name{catnl}
\alias{catnl}
\title{Cat with trailing newline.}
\description{\code{\link{cat}} with trailing newline.}
\usage{catnl(..., sep = " ")}
\arguments{
  \item{\dots}{arguments to cat}
  \item{sep}{
}
}

\value{None.}

\author{Tomas Sieger <tomas.sieger@seznam.cz> [aut, cre]}




\examples{
catnl('Hello, world!')
}
