\name{corCritVal}
\alias{corCritVal}
\title{Critical values for correlation coefficients.}
\description{TODO: details, reference}
\usage{corCritVal(n, alpha = 0.05, method = c("pearson", "spearman"), 
    approximate = FALSE)}
\arguments{
  \item{n}{sample size}
  \item{alpha}{significance level}
  \item{method}{
}
  \item{approximate}{
}
}



\author{Tomas Sieger}




\examples{
corCritVal(20)
}
