\name{pairs.default}
\alias{pairs.default}
\title{pairs default}

\usage{\method{pairs}{default}(x, labels, panel = points, ..., lower.panel = panel, 
    upper.panel = panel, diag.panel = NULL, text.panel = textPanel, 
    label.pos = 0.5 + has.diag/3, line.main = 3, cex.labels = NULL, 
    font.labels = 1, row1attop = TRUE, gap = 1, log = "", i1 = NULL, 
    i2 = NULL)}
\arguments{
  \item{x}{
}
  \item{labels}{
}
  \item{panel}{
}
  \item{\dots}{
}
  \item{lower.panel}{
}
  \item{upper.panel}{
}
  \item{diag.panel}{
}
  \item{text.panel}{
}
  \item{label.pos}{
}
  \item{line.main}{
}
  \item{cex.labels}{
}
  \item{font.labels}{
}
  \item{row1attop}{((begining of TS's modifications))}
  \item{gap}{
}
  \item{log}{
}
  \item{i1}{((end of TS's modifications))}
  \item{i2}{
}
}



\author{Tomas Sieger}





