\name{pairs.formula}
\alias{pairs.formula}
\title{pairs formula}

\usage{\method{pairs}{formula}(formula, data = NULL, ..., subset, na.action = stats::na.pass)}
\arguments{
  \item{formula}{
}
  \item{data}{
}
  \item{\dots}{
}
  \item{subset}{
}
  \item{na.action}{
}
}



\author{Tomas Sieger}





