\name{plot.default}
\alias{plot.default}
\title{'plot.default' with default plotting character 19 and no frame.}

\usage{\method{plot}{default}(..., pch = 19, frame = FALSE)}
\arguments{
  \item{\dots}{
}
  \item{pch}{
}
  \item{frame}{
}
}



\author{Tomas Sieger}




\examples{
with(iris,plot(Sepal.Length, Sepal.Width))
}
