\name{plotd}
\alias{plotd}
\title{Plot with defaults.}
\description{\code{\link{plotd}} is just a shortcut to \code{plot}
passing it some default values (no frame and large points).}
\usage{plotd(...)}
\arguments{
  \item{\dots}{parameters to \code{plot}}
}



\author{Tomas Sieger}




\examples{
plotd(1:10)
}
