\name{printWithName}
\alias{printWithName}
\alias{.pn}

\title{Print the name and value of a variable.}
\description{'printWithName' prints both the name and value of a variable. Useful
for debug purposes.
}
\usage{printWithName(x)}
\arguments{
  \item{x}{variable}
}



\author{Tomas Sieger}



\seealso{'NCmisc::preview', 'NCmisc::prv'}
\examples{
a<-1:10
.pn(a)
}
