\name{tsiMisc-package}
\alias{tsiMisc-package}
\alias{tsiMisc}
\docType{package}
\title{Miscelaneous utility functions, mostly related to exploratory statistical analysis.}
\description{Miscelanous utility functions, mostly related to exploratory statistical analysis.}
\details{
\tabular{ll}{Package: \tab tsiMisc\cr
Type: \tab Package\cr
Title: \tab Miscelaneous utility functions, mostly related to exploratory statistical analysis.\cr
Version: \tab 0.0.0.9000\cr
Date: \tab 2015-05-29\cr
Author: \tab Tomas Sieger\cr
Depends: \tab rgl\cr
Imports: \tab devtools, e1071, gdata, geometry, inlinedocs, stringr\cr
Suggests: \tab idendro, testthat, coin, MASS, scales\cr
URL: \tab http://github.com/tsieger/tsiMisc\cr
Maintainer: \tab Tomas Sieger <tomas.sieger@seznam.cz>\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr}
}
\author{Tomas Sieger}

\keyword{ package }


