\name{buildInlineDoc}
\alias{buildInlineDoc}
\title{inlinedocs build}
\description{This is simple a delegate wrapper for 'inlinedocs::package.skeleton.dx' to
to generate inline docs for an R package.}
\usage{buildInlineDoc(path = "~/src/Rweb/tsiMisc")}
\arguments{
  \item{path}{path to R package source}
}



\author{Tomas Sieger}




\examples{
##dontrun<<
# buildInlineDoc('~/src/R/myPackage')
##end<<
}
