\name{seqRob}
\alias{seqRob}
\title{Robust sequence generation resembling the matlab ':' operator.}

\usage{seqRob(from = 1, to = 1, by = NULL, ...)}
\arguments{
  \item{from}{from}
  \item{to}{to}
  \item{by}{increment}
  \item{\dots}{
}
}

\value{Sequence similar to the result of 'base::seq'.}

\author{Tomas Sieger}




\examples{
seqRob(from = 1, to = 0, by = 1)
}
