\name{tsiMisc-package}
\alias{tsiMisc-package}
\alias{tsiMisc}
\docType{package}
\title{Miscelaneous utility functions, mostly related to exploratory statistical analysis.}
\description{Miscelanous utility functions, mostly related to exploratory statistical analysis.}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tsiMisc}
\packageIndices{tsiMisc}
}
\author{Tomas Sieger <tomas.sieger@seznam.cz> [aut, cre]}

Maintainer: \packageMaintainer{tsiMisc}

\keyword{ package }
