\name{vectorLength}
\alias{vectorLength}
\title{Vector length.}
\description{'vectorLength' computes the length of a vector.}
\usage{vectorLength(v)}
\arguments{
  \item{v}{a vector}
}

\value{vetor length}

\author{Tomas Sieger}




\examples{
vectorLength(c(1, 0, 0))
vectorLength(c(1, 1, 0))
vectorLength(c(1, 1, 1))
vectorLength(c(-1, 2, 3))
}
